+++ models/job.py
@@ -4,10 +4,22 @@ from uuid import UUID
 from langchain.pydantic_v1 import BaseModel
 
 
+class HourlyRange(BaseModel):
+    start: float
+    end: float
+
+    def __repr__(self):
+        return f"${self.start}-${self.end}"
+
+    def __str__(self):
+        return self.__repr__()
+
+
 class Job(BaseModel):
     id: Optional[UUID]
     title: str
     description: str
+    hourly_range: Optional[HourlyRange]
     link: str
 
     def __repr__(self):
@@ -17,7 +29,7 @@ class Job(BaseModel):
         prefix = '##'
         if index is not None:
             prefix += f" {index+1}."
-        return f"{prefix} {self.title}\n\n{self.link}"
+        return f"{prefix} {self.title}\n_Hourly Range:_ {self.hourly_range}\n\n{self.link}"
 
     def detailed_repr(self) -> Iterator[str]:
         yield f"\n\n# {self.title}\n"